Update tests and logic to confirm to new modularization
