+++ tests/http/getThemesSearch.test.js
@@ -1,5 +1,5 @@
 const endpoint = require("../../src/controllers/getThemesSearch.js");
-const database = require("../../src/database.js");
+const database = require("../../src/database/_export.js");
 const context = require("../../src/context.js");
 
 const genPackage = require("../helpers/package.jest.js");
